requirements.txt
Flask==0.12.2
gunicorn==19.7.1
requests==2.18.4
jsonschema==2.6.0
numpy==1.14.1
marisa-trie==0.7.4
heapdict==1.0.0
chardet==3.0.4
click==6.7
future==0.16.0
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
line-bot-sdk==1.5.0
MarkupSafe==1.0
urllib3==1.22
Werkzeug==0.12.2
pymongo==3.6.1
Flask-PyMongo==0.5.1
